refactor: split ai logic into separate files
