+++ packages/ai-commands/src/types.ts
@@ -1,0 +1,4 @@
+export interface Message {
+  role: 'user' | 'assistant'
+  content: string
+}
